% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frc.R
\name{sCC}
\alias{sCC}
\title{Calculation of s-CC}
\usage{
sCC(x, y, B = 11, seed = 1)
}
\arguments{
\item{x}{a numeric feature vector containing one feature's values}

\item{y}{a numeric or character response vector containing class labels; its length must be equal to that of x}

\item{B}{the number of random splits, with each split dividing the observations into two halves; default value: 11}

\item{seed}{an integer of the random seed value; default value: 1}
}
\value{
sCC: a numeric value between 0 and 1, a smaller value indicating a more predictive feature
}
\description{
This function loads a feature x (numeric, containing feature values)
and a response vector (numeric or character, containing class labels).
}
