% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frc.R
\name{sNPC}
\alias{sNPC}
\title{Calculation of s-NPC}
\usage{
sNPC(x, y, imp_class = sort(y)[1], alpha = 0.3, delta = 0.05, B = 11, seed = 1)
}
\arguments{
\item{x}{a numeric feature vector containing one feature's values}

\item{y}{a numeric or character response vector containing class labels; its length must be equal to that of x}

\item{imp_class}{a numeric or character value of the most important class; the value must be in y; the most important class would be treated as class 0, and the other classes would be combined into class 1; default value: the first value of y in numeric or alphabetical order}

\item{alpha}{a numeric value between 0 and 1, the upper bound on the population type I error (i.e., the probability that a class 0 observation is predicted as class 1); default value: 0.3}

\item{delta}{a numeric value between 0 and 1, the violation rate of the population type I error (i.e., the probability that the population type I error exceeds alpha); default value: 0.05}

\item{B}{the number of random splits, with each split dividing the observations into two halves; default value: 11}

\item{seed}{an integer of the random seed value; default value: 1}
}
\value{
sNPC: a numeric value between 0 and 1, a smaller value indicating a more predictive feature
}
\description{
This function loads a feature x (numeric, containing feature values)
and a response vector (numeric or character, containing class labels).
}
